Welcome back µbot!
This is the first commit after significantly revamping the old µbot
code. There are many improvements to the internal structure, the dbus
interface and the helpers library. There is also a shiny new web
interface based on django and jquery and a perl library for writing
helpers.

The TODO file lists all the major TODO items :)